+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All changes to the MarkLogic (backend) portion of LUX capable of impacting the r
 ## Unreleased
 
 ### Added
+- Added agent recordType (Person or Group) to the user facing config (so it shows up in advanced search) ([#75](https://github.com/project-lux/lux-marklogic/issues/75))
 
 ### Changed
 - Changed via search facets to order by number of results descending ([#57](https://github.com/project-lux/lux-marklogic/issues/57))